pyproject.toml
[project]
name = "confluence-page-migration"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "atlassian-python-api>=4.0.7",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
]
